feat(backend): move enums
